Added functionality to dynamically get and display products from the database.
 Added endpoint to get product images from AWS s3 bucket.

 Scaled product images appropriately for the page.

+++ sam-with-postgres/src/ShopRepository/Models/Stock.cs
@@ -10,8 +10,7 @@ public class Stock
     [DynamoDBProperty] public string? StripeId { get; set; }
 
     [DynamoDBProperty] public string? Name { get; set; }
-
-    [DynamoDBProperty] public int? TotalStock { get; set; }
+    [DynamoDBProperty] public long? TotalStock { get; set; }
 
     [DynamoDBProperty] public string PhotoUri { get; set; } = "";
 }
